Bug resolved surrounding branching traversal
The implementation of the visitor pattern by pycparser does
not allow for further traversal of the node that is visited.
Meaning: If we visit a Compound, if that Compound has more compounds
within it -- they will not be identified by the visit method that was
already defined.

This was resolved by adding a new line to call visit on every child node
of compound.

Another issue resolved here was not seeing for-loops within if-statements.
I believe this also stemmed from the same issue, where something upstream was
visited, but then further nodes (which were if nodes) were left unvisited.
The class and method are here just to serve as a defined approach for visiting
the different components of an if statmeent (as the if-statement itself is not a compound,
but the iffalse and ifftrue MAY be compounds).