WIP cfgbuilder rewrite
Currently the node/statement types that have be worked on are:
1. Assignment
2. Decl
3. If
4. While
5. Break

Many of these are largely unchanged (simple renaming of condition
to check the node type, because pycparser uses a different naming
for the same idea), but there were some with *slightly* more editing.
This slight amount of editing was surrounding getting rid of things that
aren't present in C such as augassign and the handling for that, as well
as changing the attribute nodes that are further analyzed or used. Minor
edits in all -- but these have not been tested fully yet on an AST, explaining
the WIP in the name, as well as not all of the nodes being implemented.