Added basic block support for compounds
Compounds were this weird case where usually a compound is only found
as a component of a variable assignmet node. This isn't always the case
though, and sometimes a compound holds a collection of other nodes.

I decided to base the handling on "if" nodes, but remove anything
pertaining to conditionals, and the else component of the node. The
idea was sort of like "an If statement has a conditional check and then
a collection of statements + an else branch. What if we ignore the else
branch (there is no else branch in a compound), and then don't make any
changes to the basic block that make it look like a conditional.

The only thing that I want to make sure of later is that there is not
extra/unnecessaryblocks being created.