Added support for CFG Generation for C files
There are still some artifacts of the previous (python) analysis, meaning
that some nodes that are in the big if-else are actually never used.

Although support for C exists now, there is still limited support for labels/gotos.
Using a goto with label destination is supported, but goto with a variable destination is
not supported. Essentially there's not an easy way to differentiate a variable name with a label name,
but it is derived by recording the block number associated with a label and then at the end (after visiting
a goto), checking if the goto destination has a block associated with it (checks if the destination is a label).